requirements.txt
Rabsl-py==1.4.0
aiofiles==22.1.0
aiosqlite==0.18.0
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
attrs==22.1.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
blinker==1.6.3
cachetools==5.3.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.7
colorama==0.4.6
comm==0.1.2
contourpy==1.1.1
cycler==0.12.1
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
dm-tree==0.1.8
entrypoints==0.4
executing==1.2.0
fastapi==0.104.0
fastjsonschema==2.16.2
filelock==3.12.4
Flask==3.0.0
flatbuffers==23.5.26
fonttools==4.43.1
fqdn==1.5.1
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.14.0
h5py==3.10.0
htmlmin==0.1.12
idna==3.4
ImageHash==4.3.1
imbalanced-learn==0.11.0
imblearn==0.0
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jax==0.4.19
jedi==0.18.2
Jinja2==3.1.2
joblib==1.1.1
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.5.0
jupyter-ydoc==0.2.3
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.2
jupyter_server_ydoc==0.6.1
jupyterlab==3.6.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.20.0
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
lxml==4.9.3
Markdown==3.5
MarkupSafe==2.1.1
matplotlib==3.8.0
matplotlib-inline==0.1.6
missingno==0.5.2
mistune==2.0.4
ml-dtypes==0.2.0
multimethod==1.10
mysql==0.0.3
mysql-connector==2.2.9
mysql-connector-python==8.2.0
mysqlclient==2.2.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
networkx==3.2.1
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==22.0
pandas==1.5.2
pandas-datareader==0.10.0
pandas-profiling==3.2.0
pandocfilters==1.5.0
parso==0.8.3
phik==0.12.3
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==4.21.12
psutil==5.9.4
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic-settings==2.1.0
pydantic_core==2.10.1
Pygments==2.13.0
pyparsing==3.1.1
pyrsistent==0.19.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.4
python-multipart==0.0.6
pytz==2022.6
PyWavelets==1.5.0
pywin32==305
pywinpty==2.0.9
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.1
QtPy==2.3.0
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.11.3
seaborn==0.13.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
starlette==0.27.0
tangled-up-in-unicode==0.2.0
tensorboard==2.15.1
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-model-optimization==0.7.5
tensorflow-serving-api==2.5.0
termcolor==2.3.0
terminado==0.17.1
threadpoolctl==3.2.0
tinycss2==1.2.1
tk==0.1.0
tornado==6.2
tqdm==4.66.1
traitlets==5.7.0
typeguard==2.13.3
typing_extensions==4.8.0
uri-template==1.2.0
urllib3==1.26.15
uvicorn==0.23.2
virtualenv==20.24.5
visions==0.7.4
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==3.0.0
widgetsnbextension==4.0.5
wrapt==1.14.1
y-py==0.5.9
ypy-websocket==0.8.2
